refactor: 약관 정보 파일 분리 (#134)
